injector: Load stage2 library from embedded section
Instead of loading the stage2 library from file, embed it inside the
executable file during build. This is done using make for now due to cargo
limitations. At runtime, the currently running executable is loaded and
a section named `.bakatsugi` is copied to a memfd, which is passed to
the target.